M0: extract xor-delta crate (gorilla + timestamp pure logic)
Second step of the M0 codec workspace migration. Lifts the entire
XOR-delta (Gorilla) codec — combined chunk encode/decode, values-only,
timestamps-only, and block-statistics computation — into a new
o11y-codec-rt-xor-delta crate. Reduces o11ytsdb/rust/src/gorilla.rs
and timestamp.rs to thin extern "C" shims that marshal raw WASM
pointers into slices and dispatch.

What moved:
- o11y-codec-rt/xor-delta/Cargo.toml — new workspace member.
- o11y-codec-rt/xor-delta/src/lib.rs — pure-Rust slice API:
  encode_chunk / decode_chunk (combined timestamps + values),
  encode_values / decode_values (values-only),
  encode_timestamps / decode_timestamps (timestamps-only),
  compute_stats (block stats).
  The 4-tier delta-of-delta prefix coder — previously duplicated
  inline in encodeChunk and encodeTimestamps — lifted into shared
  write_dod/read_dod helpers.
- 24 unit tests moved with the logic.

What stayed in o11ytsdb/rust/:
- gorilla.rs: 6 extern "C" shims (encodeChunk, decodeChunk,
  encodeChunkWithStats, encodeValues, decodeValues, encodeValuesWithStats)
  that take raw pointers, slice them, and call into the workspace
  crate. Was 615 lines; now ~100.
- timestamp.rs: 2 extern "C" shims (encodeTimestamps, decodeTimestamps).
  Was 237 lines; now ~35.

Internal consumers updated:
- batch.rs imports compute_stats / encode_values / decode_values from
  the new crate.
- delta_alp.rs imports compute_stats from the new crate.
- range_decode.rs imports decode_timestamps from the new crate.
- verification.rs imports compute_stats from the new crate.

Verification:
- 19 codec-rt-core unit tests pass.
- 24 codec-rt-xor-delta unit tests pass (the gorilla + timestamp
  test suites lifted intact).
- 68 o11ytsdb-wasm cargo tests pass (was 91; the 23 missing live in
  the new crate now).
- 561 TS tests pass repo-wide.
- typecheck + biome lint clean.

Wasm size delta vs PR #183: -120 raw / -2 gz (the lifted
write_dod/read_dod helpers consolidate code that used to be
duplicated between gorilla and timestamp). Cumulative delta vs
pre-M0 baseline: +196 raw (+0.009%) / +268 gz (+1.48%). PLAN.md
updated with the cumulative table.

Follow-ups:
- Extract alp (alp.rs + delta_alp.rs + alp_exc.rs).
- Extract fastlanes-bp (range_decode.rs's bit-packing parts).
- Reduce o11ytsdb/rust/src/lib.rs to a thin extern "C" binding layer
  once everything's lifted.
