+++ packages/o11y-codec-rt/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["core"]
+members = ["core", "xor-delta"]
 
 # Shared codec runtime for o11ykit engines. Each engine's binding crate
 # (packages/o11ytsdb/rust, packages/o11ylogsdb/rust, …) depends on the
