+++ packages/o11y-codec-rt/xor-delta/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "o11y-codec-rt-xor-delta"
+version = "0.0.1"
+edition.workspace = true
+license.workspace = true
+publish.workspace = true
+
+# XOR-delta (Gorilla) codec for f64 values + delta-of-delta timestamps.
+# Pure-Rust slice-in / slice-out API; the WASM extern "C" surface lives
+# in the consuming engine's binding crate.
+#
+# Reference: Pelkonen et al., VLDB 2015.
+
+[lib]
+
+[dependencies]
+o11y-codec-rt-core = { path = "../core" }
+
+[lints]
+workspace = true
